type fix

+++ crates/l2/ethereum-instance/src/lib.rs
@@ -152,7 +152,7 @@ pub async fn deploy_contract<T: Tokenize>(
     Ok(factory
         .deploy(contructor_args)
         .map_err(Into::<ContractError<LocalWalletSignerMiddleware>>::into)?
-        .confirmations(2)
+        .confirmations(2 as usize)
         .send()
         .await
         .map_err(Into::<ContractError<LocalWalletSignerMiddleware>>::into)?)
